rename make files in wql
